Test Plan added, Design Document updated for Sprint 3. Added Unit Testing and Integration Testing for users, added custom validation exception handler, added Service Validations.

+++ football-tickets-app/src/main/java/com/ultras/footbalticketsapp/model/Role.java
@@ -14,8 +14,9 @@ import javax.persistence.*;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Role {
+    //TODO remove class since user uses enum
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
     private String name;
 }
